WIP: Move Condition para boosters Refatora Condition para ter StatsBooster como herança Adiciona teste para o ConditionModel
