+++ repository/mongo/models/condition.py
@@ -1,6 +1,6 @@
 from repository.mongo import Model
 from repository.mongo.collection_enum import CollectionEnum
-from rpgram import Condition
+from rpgram.boosters import Condition
 
 
 class ConditionModel(Model):
@@ -10,4 +10,16 @@ class ConditionModel(Model):
 
 
 if __name__ == '__main__':
-    ...
+    condition = Condition(
+        name='BURN TESTE',
+        description='Queimaduras que reduzem a Constituição em 10%.',
+        function='self.__multiplier_constitution = 0.9',
+        battle_function='self.__multiplier_constitution = 0.9',
+        _id='ffffffffffffffffffffffff'
+    )
+    print('CONDITION:', condition)
+    conditions_model = ConditionModel()
+    result = conditions_model.save(condition)
+    print('result:', result)
+    condition2 = conditions_model.get('ffffffffffffffffffffffff')
+    print('CONDITION2:', condition2)
