+++ rpgram/boosters/__init__.py
@@ -2,3 +2,4 @@ from rpgram.boosters.stats_booster import StatsBooster
 from rpgram.boosters.race import Race
 from rpgram.boosters.classe import Classe
 from rpgram.boosters.equipment import Equipment
+from rpgram.boosters.condition import Condition
